% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_selection_BayTetra.R
\name{Model_selection_BayTetra}
\alias{Model_selection_BayTetra}
\title{Model Selection for BayTetra}
\usage{
Model_selection_BayTetra(
  df_min,
  df_max,
  data,
  v_rsp,
  v_covs,
  v_grp,
  v_time,
  mcmc = list(),
  prior = list(),
  display_process = TRUE
)
}
\arguments{
\item{df_min}{Minimum value for degrees of freedom (df). Must be 3 or greater. See \code{mcmc_BayTetra} documentation.}

\item{df_max}{Maximum value for degrees of freedom (df).}

\item{data}{See \code{mcmc_BayTetra} documentation.}

\item{v_rsp}{See \code{mcmc_BayTetra} documentation.}

\item{v_covs}{See \code{mcmc_BayTetra} documentation.}

\item{v_grp}{See \code{mcmc_BayTetra} documentation.}

\item{v_time}{See \code{mcmc_BayTetra} documentation.}

\item{mcmc}{See \code{mcmc_BayTetra} documentation.}

\item{prior}{See \code{mcmc_BayTetra} documentation.}

\item{display_process}{See \code{mcmc_BayTetra} documentation.}
}
\value{
A list containing two elements:
\itemize{
\item{\code{selection_elpd}: A list containing a named vector of elpd values,
which elements' names are the degrees of freedom. It also contains the optimal degree of freedom.}
\item{\code{best_mcmc_result}: The best model's MCMC posterior samples (model with the highest elpd).}
}
}
\description{
Function implements the model selection of BayTetra for the degree of freedom in B-splines.
}
\examples{
\dontrun{
selection_result = Model_selection_BayTetra(
                         df_min = 4, df_max = 6,
                         data = ex_data,
                         v_rsp = paste("R", 1:2, sep = ""),
                         v_covs = "cov1",
                         v_grp = "Group",
                         v_time = "time"
                         )
}

}
\seealso{
\code{\link[=mcmc_BayTetra]{mcmc_BayTetra()}}
}
